Fixes Sum Function
Removed REST parameters so that the function works properly. It wasn't working before because REST parameters are in an array. Would need to create a helper function to check if it's an array and then use the SPREAD syntax to take the items out of the array and then the function would work.